requirements.txt
langchain>=0.0.230
openai>=0.27.0
chromadb>=0.3.21
tiktoken>=0.4.0
yfinance>=0.2.18
pandas>=1.3.0
numpy>=1.21.0
streamlit>=1.20.0
speechrecognition>=3.8.1
gtts>=2.3.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
